+++ setup.cfg
@@ -100,6 +100,7 @@ markers =
 	api: magpie API operations
 	cli: magpie CLI helper operations
 	ui: magpie UI operations
+	webhooks: magpie webhooks operations
 	utils: magpie utility functions
 	functional: magpie functional operations
 	auth_admin: magpie operations that require admin-level access
